+++ src/main/java/de/jformchecker/elements/SelectInput.java
@@ -1,6 +1,8 @@
 package de.jformchecker.elements;
 
+import java.util.ArrayList;
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
 
 import de.jformchecker.FormCheckerElement;
@@ -8,12 +10,8 @@ import de.jformchecker.TagAttributes;
 
 public class SelectInput extends AbstractInput<SelectInput> implements FormCheckerElement {
 
-	LinkedHashMap<String, String> possibleNames = new LinkedHashMap<>(); // Linked
-																			// Hashmap
-																			// to
-																			// maintain
-																			// sort
-																			// order
+	List<SelectInputEntry> entries = new ArrayList<>();
+	
 
 	public static SelectInput build(String name) {
 		SelectInput i = new SelectInput();
@@ -21,7 +19,8 @@ public class SelectInput extends AbstractInput<SelectInput> implements FormCheck
 		return i;
 	}
 
-	// convenience method
+	// a map is not a good structure for that, because keys may be used more than once
+	@Deprecated
 	public static SelectInput build(String name, LinkedHashMap<String, String> possibleNames) {
 		SelectInput si = SelectInput.build(name);
 		si.setPossibleValues(possibleNames);
@@ -34,11 +33,11 @@ public class SelectInput extends AbstractInput<SelectInput> implements FormCheck
 		if (keys.length != values.length) {
 			throw new IllegalArgumentException("Key / Values with unequal length");
 		}
-		LinkedHashMap<String, String> possibleNames = new LinkedHashMap<>();
+		List<SelectInputEntry> entries = new ArrayList<>();
 		for (int i = 0; i < keys.length; i++) {
-			possibleNames.put(keys[i], values[i]);
+			entries.add(si.generateEntry(keys[i], values[i]));
 		}
-		si.setPossibleValues(possibleNames);
+		si.setEntries(entries);
 		return si;
 	}
 
@@ -47,25 +46,39 @@ public class SelectInput extends AbstractInput<SelectInput> implements FormCheck
 		TagAttributes tagAttributes = new TagAttributes(attributes);
 		tagAttributes.add(this.inputAttributes);
 		StringBuilder inputTag = new StringBuilder("<select " + buildAllAttributes(tagAttributes) + " name=\"" + name + "\" >\n");
-		possibleNames.forEach((key,v)->{
+		entries.forEach(entry->{
 			String sel = "";
-			if (value != null && value.equals(key)) {
+			if (value != null && value.equals(entry.key)) {
 				sel = " SELECTED ";
 			}
-			inputTag.append("<option value=\"" + key + "\"" + sel + ">" + possibleNames.get(key) + "</option>\n");
+			inputTag.append("<option value=\"" + entry.key + "\"" + sel + ">" + entry.value + "</option>\n");
 		});
 		return inputTag.append("</select>\n").toString();
 	}
 
+	@Deprecated
 	public SelectInput setPossibleValues(LinkedHashMap<String, String> possibleNames) {
-		this.possibleNames = possibleNames;
+		List<SelectInputEntry> entries = new ArrayList<>();
+		possibleNames.forEach((k,v) ->
+			entries.add(this.generateEntry(k, v))
+				);
+		this.entries = entries;
 		return this;
 	}
+	
+	public SelectInput setEntries(List<SelectInputEntry> entries) {
+		this.entries = entries;
+		return this;
+	}
+	
 
 	@Override
 	public void setValue(String value) {
-		if (possibleNames.containsKey(value)) {
-			this.value = value;
+		for (SelectInputEntry entry : entries) {
+			if (value != null && value.equals(entry.getKey())){
+				this.value = value;
+				break;
+			}
 		}
 	}
 	
@@ -74,5 +87,26 @@ public class SelectInput extends AbstractInput<SelectInput> implements FormCheck
 		return "select";
 	}
 
+	public SelectInputEntry generateEntry(String key, String value) {
+		return new SelectInputEntry(key, value);
+	}
+	
+	// class that represents an entry in the selectInput
+	public class SelectInputEntry {
+		private final String key;
+		private final String value;
+		public SelectInputEntry(String key, String value) {
+			this.key = key;
+			this.value = value;
+		}
+		public String getKey() {
+			return key;
+		}
+		public String getValue() {
+			return value;
+		}
+		
+		
+	}
 
 }